Added a jupyter notebook file
All training, testing and evaluation of machine learning models applied to the UCI Poker Hand Dataset is contained within this file. The models applied are Support Vector Machine, K-nearest Neighbor, Random Forest and finally, a sequential model. 

The comments in the code clearly describe the various challenges and steps taken as proceeded with the dataset.